# NowqttGateway/spec/swagger.yaml
openapi: 3.0.3

info:
  version: '2.2.0'
  title: Nowqtt Gateway API
  description: Inspect the Nowqtt mesh, manage device names, and control OTA updates.

servers:
  - url: /v1

tags:
  - name: Runtime
  - name: Traces
  - name: Devices
  - name: OTA

paths:
  /health:
    get:
      summary: Get gateway health and runtime metrics
      tags: [Runtime]
      responses:
        '200':
          description: All required workers, MQTT, and serial I/O are healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          description: At least one required runtime component is unhealthy or stopping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /traces:
    get:
      summary: Get recent route traces
      tags: [Traces]
      parameters:
        - $ref: '#/components/parameters/DeviceMacQuery'
        - $ref: '#/components/parameters/Last'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A bounded page of traces.
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trace'
        '400':
          $ref: '#/components/responses/BadRequest'

  /devices:
    get:
      summary: Get devices observed in route traces
      tags: [Devices]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A bounded page of devices.
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'

  /devices/names:
    get:
      summary: Get stored device names
      tags: [Devices]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A bounded page of device names.
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceName'
        '400':
          $ref: '#/components/responses/BadRequest'

  /devices/{device_mac_address}/names:
    parameters:
      - $ref: '#/components/parameters/DeviceMacPath'
    get:
      summary: Get the stored name for one device
      tags: [Devices]
      responses:
        '200':
          description: The matching name collection, which may be empty.
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceName'
        '400':
          $ref: '#/components/responses/BadRequest'
    patch:
      summary: Set a manual device name
      tags: [Devices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        '201':
          description: The stored device name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceName'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      summary: Remove a stored device name
      tags: [Devices]
      responses:
        '204':
          description: The name was removed or did not exist.
        '400':
          $ref: '#/components/responses/BadRequest'

  /devices/activity:
    get:
      summary: Get device availability activity
      tags: [Devices]
      parameters:
        - $ref: '#/components/parameters/DeviceMacQuery'
        - $ref: '#/components/parameters/Last'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A bounded page of activity records and current totals.
          content:
            application/json:
              schema:
                type: object
                required: [total, online, offline, items]
                properties:
                  total:
                    type: integer
                  online:
                    type: integer
                  offline:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceActivity'
        '400':
          $ref: '#/components/responses/BadRequest'

  /ota/update/{device_mac_address}:
    parameters:
      - $ref: '#/components/parameters/DeviceMacPath'
    post:
      summary: Start an OTA update
      tags: [OTA]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Non-empty ESP firmware with a .bin filename.
      responses:
        '202':
          description: The OTA session was accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtaStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Another OTA session is already active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: The upload exceeds the configured firmware size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ota/status:
    get:
      summary: Get active or most recent OTA status
      tags: [OTA]
      responses:
        '200':
          description: OTA coordinator status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtaStatus'

components:
  parameters:
    DeviceMacPath:
      name: device_mac_address
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{12}$'
    DeviceMacQuery:
      name: device_mac_address
      in: query
      required: false
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{12}$'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Last:
      name: last
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0

  responses:
    BadRequest:
      description: Invalid path, query, JSON, or upload data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Device:
      type: object
      required: [mac_address]
      properties:
        mac_address:
          type: string
          pattern: '^[0-9a-f]{12}$'
        name:
          type: string
          nullable: true
    DeviceName:
      type: object
      required: [name, mac_address]
      properties:
        name:
          type: string
        mac_address:
          type: string
          pattern: '^[0-9a-f]{12}$'
    DeviceActivity:
      type: object
      required: [mac_address, timestamp, activity]
      properties:
        name:
          type: string
          nullable: true
        mac_address:
          type: string
        timestamp:
          type: string
        activity:
          type: integer
          enum: [0, 1]
    Hop:
      type: object
      properties:
        hop_counter:
          type: integer
        hop_mac_address:
          type: string
        name:
          type: string
          nullable: true
        hop_rssi:
          type: integer
        hop_dest_seq:
          type: integer
        route_age:
          type: integer
        hop_count:
          type: integer
    Trace:
      type: object
      required: [uuid, dest_mac_address, timestamp, hops]
      properties:
        uuid:
          type: string
          format: uuid
        dest_mac_address:
          type: string
        name:
          type: string
          nullable: true
        timestamp:
          type: string
        hops:
          type: array
          items:
            $ref: '#/components/schemas/Hop'
    OtaStatus:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
        device_mac_address:
          type: string
        state:
          type: string
        packet_count:
          type: integer
        sent_packets:
          type: integer
    Health:
      type: object
      required:
        - healthy
        - stopping
        - mqtt_connected
        - serial_reader
        - serial_writer
        - runtime_workers
        - ota
      properties:
        healthy:
          type: boolean
        stopping:
          type: boolean
        mqtt_connected:
          type: boolean
        serial_reader:
          type: object
          additionalProperties: true
        serial_writer:
          type: object
          additionalProperties: true
        runtime_workers:
          type: object
          additionalProperties: true
        ota:
          $ref: '#/components/schemas/OtaStatus'